+++ example_getx/lib/src/pages/main_page.dart
@@ -7,7 +7,7 @@ import 'package:example_getx/example_getx_routes.dart' as example_routes;
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:url_launcher/url_launcher_string.dart';
 
 @FFRoute(
   name: 'fluttercandies://mainpage',
@@ -53,7 +53,9 @@ class MainPage extends StatelessWidget {
                 ),
               ),
               onPressed: () {
-                launch('https://github.com/fluttercandies/ff_annotation_route');
+                launchUrlString(
+                  'https://github.com/fluttercandies/ff_annotation_route',
+                );
               },
             ),
           ),
@@ -64,7 +66,9 @@ class MainPage extends StatelessWidget {
               child:
                   Image.network('https://pub.idqqimg.com/wpa/images/group.png'),
               onPressed: () {
-                launch('https://jq.qq.com/?_wv=1027&k=5bcc0gy');
+                launchUrlString(
+                  'https://jq.qq.com/?_wv=1027&k=5bcc0gy',
+                );
               },
             ),
           )