+++ example_route_extension/android/app/src/main/kotlin/com/example/example_route_extension/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.example_route_extension
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}